# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "REST API for a Laravel-based news aggregator service.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get all articles",
                "description": "Returns a list of articles with optional filters.",
                "operationId": "3fe8bf684063a5479fb27f5432db77e2",
                "parameters": [
                    {
                        "name": "author",
                        "in": "query",
                        "description": "Filter by author ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter by publication date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search articles by title, slug, or content",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ArticleResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Create a new article",
                "description": "Creates a new article with the provided data.",
                "operationId": "7541a17ad3825bab482f62fbb79e6507",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArticleResource"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get article by ID",
                "description": "Returns a single article by its ID.",
                "operationId": "0e00aae7c370b295281ca65f1d061354",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Article not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/preferences": {
            "get": {
                "tags": [
                    "Preferences"
                ],
                "summary": "Save user news preferences",
                "description": "Stores the user's preferred category and news source for personalized feeds.",
                "operationId": "94ae0966450c96135b23274e389b9ce4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category",
                                    "source"
                                ],
                                "properties": {
                                    "category": {
                                        "type": "string",
                                        "example": "technology"
                                    },
                                    "source": {
                                        "type": "string",
                                        "example": "News API"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preference successfully saved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "category": {
                                                    "type": "string",
                                                    "example": "technology"
                                                },
                                                "source": {
                                                    "type": "string",
                                                    "example": "News API"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/preferences": {
            "post": {
                "tags": [
                    "Preferences"
                ],
                "summary": "Save user preferences",
                "description": "Store a newly created resource in storage.",
                "operationId": "d03dcea4a5a3eefea8e4407de1a92d2c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category": {
                                        "type": "string",
                                        "example": "technology"
                                    },
                                    "source": {
                                        "type": "string",
                                        "example": "News API"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preference stored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "category": {
                                                    "type": "string"
                                                },
                                                "source": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ArticleResource": {
                "title": "Article Resource",
                "description": "Represents an article in the news aggregator system.",
                "required": [
                    "id",
                    "title",
                    "slug",
                    "content",
                    "author",
                    "source",
                    "news_source",
                    "url",
                    "published_at"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the actor type",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Name of the actor type",
                        "type": "string"
                    },
                    "slug": {
                        "description": "Name of the actor type",
                        "type": "string"
                    },
                    "content": {
                        "description": "Name of the actor type",
                        "type": "string"
                    },
                    "author": {
                        "description": "Name of the actor type",
                        "type": "string"
                    },
                    "source": {
                        "description": "Name of the actor type",
                        "type": "string"
                    },
                    "news_source": {
                        "description": "Name of the actor type",
                        "type": "string"
                    },
                    "url": {
                        "description": "Name of the actor type",
                        "type": "string"
                    },
                    "published_at": {
                        "description": "Name of the actor type",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ArticleResource"
                }
            }
        }
    },
...
}